Refactor connection logic to separate class
